requirements.txt
Django
django-tailwind
django-tailwind[reload]
djangorestframework